requirements.txt
python-dotenv==0.20.0
Werkzeug==2.1.2
Flask==2.1.3
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.3
Flask-Mail==0.9.1
Flask-Dance==6.1.0
psycopg2-binary==2.9.3
flask-restx==0.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
flask-cors==3.0.10
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
tweepy==4.10.0
typing-extensions==4.3.0
boto3==1.24.57
Pillow==9.2.0
stripe==4.2.0